refactored server.js and included .env dependencies and file
