+++ data/dbConfig.js
@@ -1,4 +1,7 @@
 const knex = require('knex');
-const knexConfig = require('../knexfile.js');
 
-module.exports = knex(knexConfig.development);
+const knexConfig = require('../knexfile.js')[db];
+
+const db = process.env.DB_ENV || 'development';
+
+module.exports = knex(knexConfig);
