+++ package.json
@@ -24,6 +24,7 @@
   "homepage": "https://github.com/LambdaSchool/Node-Blog#readme",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "knex": "^0.16.3",
     "sqlite3": "^4.0.6"
